The goal right now is to get the FSA to report the number of characters that it consumed from the input so that the outer lexer can: 1) decide which FSA consumes more characters, so which wins 2) work through the input token by token
Currently the FSA doesn't stop running until it hits the end of the input. After looking at how this is coded it seemed simpler to count up characters consumed rather than figure out a different way to stop running. Not that it would be that hard to stop. It would actually be more efficient to stop immediately on an unacceptable state. Maybe that's a good enough reason to change the stop logic....

I also changed the name of the FSA to the name of the token so the lexer can just iterate the FSAs, determine which wins, and record off the name of it.