This is an FSA for the UNDEFINED token. This has specific rules according to the spec, and so it needs its own FSA.
The rule is that the first character is an invalid first character for any of the other tokens or it's an unterminated string. The strategy is to mostly duplicate the string FSA, but change the accept states so that if it's matching a string token the FSA only accepts if the string is in an unterminated state.

For the invalid first character state it just checks for an invalid first character in S0 and if it finds one, that's it.

My read of the spec is that this is either a single incorrect start character or it's a whole unterminated string. However, the tests are looking for an unterminated string to just have "'" as the value of the token, so I handled that by overriding the get_consumed_input function to just return one character of the consumed input.