This is an FSA for an ID. The basic rule here is that the token has to start with an alphabetical character and it includes all following alphanumeric characters. Once the input isn't alphanumeric, the ID is complete.
The one trick with this is that if the ID exactly matches a keyword it isn't valid. However it's possible for the id to start with the keyword but continue on. So when the value of this FSA matches a keyword it needs to go into a state that is not an accept state just in case that's the end of the input, but the state can continue to an accept state if there are additional characters that attach to the value. This special reserved word state is S_reserved. The behavior of the state is exactly the same as S1, the only difference is that it isn't listed as an accept state. So S_reserved simply calls S1.