Add all of the new FSAs to the lexer. I'm processing them in alphabetical order except for the undefined FSA. If for some reason both a valid FSA and the undefined FSA accept, the valid FSA should win because the undefined FSA should be a single character and won't replace the existing "best".
Take out the error handling that inserts the UNDEFINED token. The undefined FSA will do this now.

When adding tokens to the output, ignore WHITESPACE tokens. That FSA is just there to eat up white space.

If there is an UNDEFINED token we are supposed to stop outputting tokens and change the total message to state there was an error on the line.